% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libminerSA-package.R
\docType{package}
\name{libminerSA-package}
\alias{libminerSA}
\alias{libminerSA-package}
\title{libminerSA: Explore Your R Libraries}
\description{
rovides functions for learning about your R libraries, and the packages you have installed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selcant/libminerSA}
  \item \url{https://selcant.github.io/libminerSA/}
  \item Report bugs at \url{https://github.com/selcant/libminerSA/issues}
}

}
\author{
\strong{Maintainer}: Selcan Aydin \email{selcant@gmail.com} (\href{https://orcid.org/0000-0002-0457-9641}{ORCID})

}
\keyword{internal}
